docs: Elaborate on service extension
