+++ examples/demo/app/src/dog/README.md
@@ -1,0 +1,5 @@
+# Dog Service
+
+This service demontstrates the usage of service extension feature. It extends
+the [Mammal](/examples/demo/app/src/mammal/) and [Walker](/examples/demo/walker/src/)
+services.
