+++ examples/demo/app/src/mammal/README.md
@@ -1,0 +1,4 @@
+# Mammal Service
+
+This service is used as a part of the example showcasing service extension feature.
+See the [Dog](/examples/demo/app/src/dog/) service for the usage.
