+++ examples/demo/wasm/tests/gclient.rs
